Move model classes into seperate packages
