+++ src/main/java/com/refactoringflow/refactoringflowbackend/model/codefile/CodeFile.java
@@ -1,5 +1,6 @@
-package com.refactoringflow.refactoringflowbackend.model;
+package com.refactoringflow.refactoringflowbackend.model.codefile;
 
+import com.refactoringflow.refactoringflowbackend.model.assignment.Assignment;
 import lombok.*;
 import org.springframework.lang.NonNull;
