+++ src/main/java/com/refactoringflow/refactoringflowbackend/service/AssignmentService.java
@@ -1,7 +1,7 @@
 package com.refactoringflow.refactoringflowbackend.service;
 
 import com.refactoringflow.refactoringflowbackend.repository.AssignmentRepository;
-import com.refactoringflow.refactoringflowbackend.model.Assignment;
+import com.refactoringflow.refactoringflowbackend.model.assignment.Assignment;
 import com.refactoringflow.refactoringflowbackend.model.user.Student;
 import org.springframework.stereotype.Service;
